requirements.txt
gitpython
humanize